# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Subscriptions Aggregation API
  version: 1.0.0
  description: REST API for CRUD operations over user subscription records and total monthly cost calculation for a selected period.
servers:
  - url: http://localhost:8080
    description: Local Docker environment
tags:
  - name: Health
  - name: Subscriptions
paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /api/v1/subscriptions:
    post:
      tags: [Subscriptions]
      summary: Create a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
            example:
              service_name: Yandex Plus
              price: 400
              user_id: 60601fee-2bf1-4721-ae6f-7636e79a0cba
              start_date: 07-2025
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [Subscriptions]
      summary: List subscriptions
      parameters:
        - $ref: '#/components/parameters/UserID'
        - $ref: '#/components/parameters/ServiceName'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSubscriptionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/subscriptions/total:
    get:
      tags: [Subscriptions]
      summary: Calculate total subscription cost for a selected period
      description: The total is calculated as monthly price multiplied by the number of months where a subscription intersects the requested period. Month boundaries are inclusive.
      parameters:
        - name: period_start
          in: query
          required: true
          schema:
            type: string
            pattern: '^(0[1-9]|1[0-2])-\d{4}$'
            example: 07-2025
        - name: period_end
          in: query
          required: true
          schema:
            type: string
            pattern: '^(0[1-9]|1[0-2])-\d{4}$'
            example: 12-2025
        - $ref: '#/components/parameters/UserID'
        - $ref: '#/components/parameters/ServiceName'
      responses:
        '200':
          description: Total cost
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotalResponse'
              example:
                total_price: 2400
                currency: RUB
                period_start: 07-2025
                period_end: 12-2025
                user_id: 60601fee-2bf1-4721-ae6f-7636e79a0cba
                service_name: Yandex Plus
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/subscriptions/{id}:
    get:
      tags: [Subscriptions]
      summary: Get a subscription by ID
      parameters:
        - $ref: '#/components/parameters/SubscriptionID'
      responses:
        '200':
          description: Subscription found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Subscriptions]
      summary: Update a subscription
      parameters:
        - $ref: '#/components/parameters/SubscriptionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriptionRequest'
      responses:
        '200':
          description: Subscription updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Subscriptions]
      summary: Delete a subscription
      parameters:
        - $ref: '#/components/parameters/SubscriptionID'
      responses:
        '204':
          description: Subscription deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    SubscriptionID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: f23a1433-8f4e-4885-8488-30a9df7a0de1
    UserID:
      name: user_id
      in: query
      required: false
      schema:
        type: string
        format: uuid
      example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
    ServiceName:
      name: service_name
      in: query
      required: false
      schema:
        type: string
      example: Yandex Plus
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Subscription not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CreateSubscriptionRequest:
      type: object
      required: [service_name, price, user_id, start_date]
      properties:
        service_name:
          type: string
          maxLength: 255
          example: Yandex Plus
        price:
          type: integer
          minimum: 1
          description: Monthly price in Russian rubles.
          example: 400
        user_id:
          type: string
          format: uuid
          example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        start_date:
          type: string
          pattern: '^(0[1-9]|1[0-2])-\d{4}$'
          example: 07-2025
        end_date:
          type: string
          nullable: true
          pattern: '^(0[1-9]|1[0-2])-\d{4}$'
          example: 12-2025
    UpdateSubscriptionRequest:
      allOf:
        - $ref: '#/components/schemas/CreateSubscriptionRequest'
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        service_name:
          type: string
          example: Yandex Plus
        price:
          type: integer
          example: 400
        user_id:
          type: string
          format: uuid
        start_date:
          type: string
          example: 07-2025
        end_date:
          type: string
          nullable: true
          example: 12-2025
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ListSubscriptionsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Subscription'
        count:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
    TotalResponse:
      type: object
      properties:
        total_price:
          type: integer
          format: int64
          example: 2400
        currency:
          type: string
          example: RUB
        period_start:
          type: string
          example: 07-2025
        period_end:
          type: string
          example: 12-2025
        user_id:
          type: string
          format: uuid
        service_name:
          type: string
          example: Yandex Plus
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: validation failed
        details:
          type: string
          example: price must be a positive integer